OAI <metadata> wrappers containing qualified record
Each record in OAI-PMH responses must be wrapped in <metadata> and
explicitly qualified with its namespace prefix. CMDI types lack the
prefix but META-SHARE resourceInfo, OLAC, and mandatory oai_dc should
have that now.

Unfortunately, xmllint seems to fail to use the external schemas. As a
workaround, there is now another locally modified schema, OAI-PMH-lax,
that skips the validation of that part of the response.