Added schema validation to several tests
A couple of tests fail to validate: Identify record has fields
in the wrong order, ListSets apparently requires at least one
set, so may need to introduce one now rather than later. Can we
have an empty set, then? That would do it, but this protocol
tends to regard empty things as errors, so expect it to fail.