Sanat: Update Matomo installation process in tasks
This commit revises the Matomo installation process in the Ansible tasks. The changes allow fetching of the latest Matomo version, checks the existing installation and updates accordingly. It also includes enhanced handling of Matomo directories and the installation process. The timer type in services has been moved for better readability.