KP-9086 Remove duplicated heliots version info
Heliots version was specified both in site.yml and in
`heliots/defaults/main.yml`, the latter of which also contained the
download URL and thus needs to be modified every time a new version is
installed. Removed the version info from here to keep all related info
in one place.